combined & updated smart contract & related pubsub
